Hice multitud de cambios tanto en los campos como en validaciones, cambié en el controlador el hecho de borrar autores, ahora hará falta que estos no tengan fichas asignadas para poder borrarse, este cambio es en la página, la base de datos sigue sin borrarse en cascada igual, en perfil con las preguntas y una notificación si no las tienes creadas y le puse un append a cualquier cantidad de inputs, acomodé vistas, puse labels, entre otros, la única cuestión es que puedes insertar el mismo autor varias veces en la misma ficha, pero esto no implica que el sistema colapse al menos, etc.

+++ app/Http/Controllers/AutorController.php
@@ -120,7 +120,12 @@ class AutorController extends Controller
      */
     public function destroy($id)
     {
-        $autor = Autor::find($id);
+        $autor = Autor::findOrFail($id);
+
+        if ($autor->ficha()->count() > 0) {
+            return redirect()->back()->with('mensaje', 'No se puede eliminar el autor porque tiene fichas asociadas.')->with('icono', 'error');
+        }
+        
         $autor->delete();
 
         return redirect()->route('admin.autores.index')
